requirements.txt
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
google-generativeai
requests==2.32.4
pandas==2.3.1
yfinance==0.2.65
supabase
python-dotenv
